requirements.txt
aniso8601==9.0.1
Babel==2.13.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
filelock==3.13.1
flake8==6.1.0
Flask==2.3.0
Flask-Babel==2.0.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
psycopg2-binary==2.9.1
pybabel==0.0.0.dev0
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pyproject-api==1.6.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
resend==0.6.0
six==1.16.0
SpeechRecognition==3.10.1
SQLAlchemy==2.0.23
tomli==2.0.1
tox==4.11.4
typing_extensions==4.8.0
urllib3==2.1.0
virtualenv==20.25.0
Werkzeug==2.3.0
yamllint==1.33.0
zipp==3.17.0